reshuffle base classes
