+++ packages/main/src/DateTimePickerPopover.hbs
@@ -19,8 +19,8 @@
 			id="{{_id}}-calendar"
 			primary-calendar-type="{{_primaryCalendarType}}"
 			format-pattern="{{_formatPattern}}"
-			timestamp="{{_calTimestamp}}"
-			.selectedDates={{_calDates}}
+			timestamp="{{_effectiveCalendarTimestamp}}"
+			.selectedDates={{_effectiveCalendarSelectedDates}}
 			.selectionMode="{{_calendarSelectionMode}}"
 			.minDate="{{minDate}}"
 			.maxDate="{{maxDate}}"
